pyproject.toml
[tool.poetry]
name = "pymusic"
version = "0.1.0"
description = ""
authors = ["MrRikkert"]

[tool.poetry.dependencies]
python = "^3.9"
pylast = "^5.1.0"
pony = "^0.7.14"
python-dotenv = "^0.19.2"
pypiwin32 = { version="^223", markers="sys_platform == 'win32'" }
pywin32 = { version="^303", markers="sys_platform == 'win32'" }
click = "^8.0.4"
pytz = "^2021.3"
pykakasi = "^2.2.1"
mixer = "^7.0.6"
loguru = "^0.6.0"
better-exceptions = "^0.3.3"
pandas = "^1.3.3"
dash = "^2.2.0"
dash-bootstrap-components = "^1.0.3"
python-dateutil = "^2.8.2"
psycopg2-binary = "^2.9.3"
dash-bootstrap-templates = "^1.0.5"
altair = "^4.2.0"
scipy = { version="^1.7.3", python=">=3.9,<3.11" }
dash-alternative-viz = {git = "https://github.com/plotly/dash-alternative-viz.git"}
pydantic = "^1.9.0"
Pillow = "^9.3.0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
pytest = "^7.0.1"
pytest-cov = "^3.0.0"
mixer = "^7.0.6"
flake8 = "^4.0.1"
mccabe = "^0.6"
requests = "^2.27.1"
pytest-xdist = "^2.5.0"
pytest-optional-tests = "^0.1.1"
pytest-dotenv = "^0.5.2"
wcwidth = "^0.2.5"
jupyter = "^1.0.0"
isort = "^5.10.1"
autoflake = "^1.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
